+++ UnitySDK/Packages/com.happs.sdk/Runtime/WebGL/HAppsWebProvider.cs
@@ -8,7 +8,7 @@ namespace HAppsSDK
 {
     public sealed class HAppsWebProvider : HAppsProvider
     {
-        public const string Version = "3.0.1";
+        public const string Version = "3.1.0";
 
         public event Action<UserData, SignatureData> AuthCompleted;
         public string Signature { get; private set; }